test(router): add failing Router regression test for hash location strategy
This situation can probably happen only when using
HashLocationStrategy and by manually changing hash that triggers a route
guard that returns a new UrlTree. Then hash in the browser might not
match the current route because navigation was canceled, while hash in
the URL remained set by the user.

Related to #37048